Add for loop through column items at line 467 to find groups of column squares that have the same possible numbers as each column item. If the number of matched column items is equal to the number of possilbe numbers, remove those possible numbers from other column squares of that column.
Add for loop through row items at line 507 to find groups of row squares that have the same possible numbers as each row item. If the number of matched row items is equal to the number of possible numbers, remove those possible numbers from other row squares of that row.